Attach mousemove events to the document
To avoid "slip-off" from the decals brace, attach mousemove and touchmove to
the document.  Only attach mousedown and touchstart to the node, but set a flag
to pick the right event handler.

Similarly, stop "slip-off" in the color picker (between shade and hue) by
resetting the bounding rectangle on mouseup/touchend.

Also reset history during reset().